Added Infinite Mode to IKG
A new Infinite mode in the settings menu allows the IKG fight to last indefinitely. The Geo counter will keep track of how much damage you have dealt. Works for both IKG and AIKG, but unsure if it works with God Mode

Added a final score display that gets shown when the player dies and leaves the battle scene

Added a performance mode that removes decor for added performance

God Mode's health is now increased by 1.75x, putting it on roughly the same level as NGG. This can be toggled off in the settings menu